Conflicts:
	packages/sysutils/systemd/package.mk

+++ packages/graphics/mesa/package.mk
@@ -17,20 +17,31 @@
 ################################################################################
 
 PKG_NAME="mesa"
-PKG_VERSION="12.0.1"
-PKG_REV="1"
+PKG_VERSION="17.2.3"
+PKG_SHA256="a0b0ec8f7b24dd044d7ab30a8c7e6d3767521e245f88d4ed5dd93315dc56f837"
 PKG_ARCH="any"
 PKG_LICENSE="OSS"
 PKG_SITE="http://www.mesa3d.org/"
-PKG_URL="ftp://freedesktop.org/pub/mesa/${PKG_VERSION%-*}/$PKG_NAME-$PKG_VERSION.tar.xz"
-PKG_DEPENDS_TARGET="toolchain Python:host expat glproto dri2proto presentproto libdrm libXext libXdamage libXfixes libXxf86vm libxcb libX11 systemd dri3proto libxshmfence libressl"
-PKG_PRIORITY="optional"
+PKG_URL="ftp://freedesktop.org/pub/mesa/$PKG_NAME-$PKG_VERSION.tar.xz"
+PKG_DEPENDS_TARGET="toolchain expat libdrm"
 PKG_SECTION="graphics"
 PKG_SHORTDESC="mesa: 3-D graphics library with OpenGL API"
 PKG_LONGDESC="Mesa is a 3-D graphics library with an API which is very similar to that of OpenGL*. To the extent that Mesa utilizes the OpenGL command syntax or state machine, it is being used with authorization from Silicon Graphics, Inc. However, the author makes no claim that Mesa is in any way a compatible replacement for OpenGL or associated with Silicon Graphics, Inc. Those who want a licensed implementation of OpenGL should contact a licensed vendor. While Mesa is not a licensed OpenGL implementation, it is currently being tested with the OpenGL conformance tests. For the current conformance status see the CONFORM file included in the Mesa distribution."
-
-PKG_IS_ADDON="no"
 PKG_AUTORECONF="yes"
+
+if [ "$DISPLAYSERVER" = "x11" ]; then
+  PKG_DEPENDS_TARGET="$PKG_DEPENDS_TARGET glproto dri2proto presentproto libXext libXdamage libXfixes libXxf86vm libxcb libX11 dri3proto libxshmfence"
+  export DRI_DRIVER_INSTALL_DIR=$XORG_PATH_DRI
+  export DRI_DRIVER_SEARCH_DIR=$XORG_PATH_DRI
+  export X11_INCLUDES=
+  MESA_DRI="--enable-dri --enable-dri3"
+  MESA_GLX="--enable-glx --enable-driglx-direct --enable-glx-tls"
+  MESA_EGL_PLATFORMS="--with-platforms=x11,drm"
+else
+  MESA_DRI="--enable-dri --disable-dri3"
+  MESA_GLX="--disable-glx --disable-driglx-direct --disable-glx-tls"
+  MESA_EGL_PLATFORMS="--with-platforms=drm"
+fi
 
 # configure GPU drivers and dependencies:
   get_graphicdrivers
@@ -38,9 +49,9 @@ PKG_AUTORECONF="yes"
 if [ "$LLVM_SUPPORT" = "yes" ]; then
   PKG_DEPENDS_TARGET="$PKG_DEPENDS_TARGET elfutils llvm"
   export LLVM_CONFIG="$SYSROOT_PREFIX/usr/bin/llvm-config-host"
-  MESA_GALLIUM_LLVM="--enable-gallium-llvm --enable-llvm-shared-libs"
+  MESA_GALLIUM_LLVM="--enable-llvm --enable-llvm-shared-libs"
 else
-  MESA_GALLIUM_LLVM="--disable-gallium-llvm"
+  MESA_GALLIUM_LLVM="--disable-llvm"
 fi
 
 if [ "$VDPAU_SUPPORT" = "yes" -a "$DISPLAYSERVER" = "x11" ]; then
@@ -56,33 +67,29 @@ for drv in $GRAPHIC_DRIVERS; do
 done
 
 if [ "$OPENGLES_SUPPORT" = "yes" ]; then
-  MESA_GLES="--enable-gles2"
+  MESA_GLES="--disable-gles1 --enable-gles2"
 else
-  MESA_GLES="--disable-gles2"
+  MESA_GLES="--disable-gles1 --disable-gles2"
 fi
- 
+
 PKG_CONFIGURE_OPTS_TARGET="CC_FOR_BUILD=$HOST_CC \
                            CXX_FOR_BUILD=$HOST_CXX \
                            CFLAGS_FOR_BUILD= \
                            CXXFLAGS_FOR_BUILD= \
                            LDFLAGS_FOR_BUILD= \
-                           X11_INCLUDES= \
-                           DRI_DRIVER_INSTALL_DIR=$XORG_PATH_DRI \
-                           DRI_DRIVER_SEARCH_DIR=$XORG_PATH_DRI \
                            --disable-debug \
                            --disable-mangling \
                            --enable-texture-float \
                            --enable-asm \
                            --disable-selinux \
                            --enable-opengl \
-                           --disable-gles1 \
                            $MESA_GLES \
-                           --enable-dri \
-                           --enable-dri3 \
-                           --enable-glx \
+                           $MESA_DRI \
+                           $MESA_GLX \
                            --disable-osmesa \
                            --disable-gallium-osmesa \
-                           --enable-egl --with-egl-platforms=x11,drm \
+                           --enable-egl \
+                           $MESA_EGL_PLATFORMS \
                            $XA_CONFIG \
                            --enable-gbm \
                            --disable-nine \
@@ -92,31 +99,31 @@ PKG_CONFIGURE_OPTS_TARGET="CC_FOR_BUILD=$HOST_CC \
                            --disable-va \
                            --disable-opencl \
                            --enable-opencl-icd \
-                           --disable-xlib-glx \
-                           --disable-r600-llvm-compiler \
                            --disable-gallium-tests \
                            --enable-shared-glapi \
                            --enable-shader-cache \
-                           --enable-sysfs \
-                           --enable-driglx-direct \
-                           --enable-glx-tls \
                            $MESA_GALLIUM_LLVM \
                            --disable-silent-rules \
                            --with-gl-lib-name=GL \
                            --with-osmesa-lib-name=OSMesa \
                            --with-gallium-drivers=$GALLIUM_DRIVERS \
                            --with-dri-drivers=$DRI_DRIVERS \
+                           --with-vulkan-drivers=no \
                            --with-sysroot=$SYSROOT_PREFIX"
 
 pre_configure_target() {
-  export LIBS="-lxcb-dri3 -lxcb-present -lxcb-sync -lxshmfence -lz"
+  if [ "$DISPLAYSERVER" = "x11" ]; then
+    export LIBS="-lxcb-dri3 -lxcb-dri2 -lxcb-xfixes -lxcb-present -lxcb-sync -lxshmfence -lz"
+  fi
 }
 
 post_makeinstall_target() {
-  # rename and relink for cooperate with nvidia drivers
+  if [ "$DISPLAYSERVER" = "x11" ]; then
+    # rename and relink for cooperate with nvidia drivers
     rm -rf $INSTALL/usr/lib/libGL.so
     rm -rf $INSTALL/usr/lib/libGL.so.1
     ln -sf libGL.so.1 $INSTALL/usr/lib/libGL.so
     ln -sf /var/lib/libGL.so $INSTALL/usr/lib/libGL.so.1
     mv $INSTALL/usr/lib/libGL.so.1.2.0 $INSTALL/usr/lib/libGL_mesa.so.1
+  fi
 }
